# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Filmes",
    "version": "1.0.0",
    "description": "API para cadastro de Filmes e visualização das informações"
  },
  "paths": {
    "/Filmes": {
      "get": {
        "summary": "Obter a lista de Filmes",
        "responses": {
          "200": {
            "description": "Lista de Filmes retornada com sucesso"
          }
        }
      },
      "post": {
        "summary": "Cadastrar um novo filme",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Filmes"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cadastro realizado com sucesso"
          },
          "400": {
            "description": "Dados incompletos"
          }
        }
      }
    },
    "/Filmes/{id}": {
      "put": {
        "summary": "Atualizar informações de um Filme pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Filmes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atualização feita com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Excluir um Filme pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "excluido"
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Filmes": {
        "type": "object",
        "properties": {
          "capa": {
            "type": "URl",
            "format": "imagem"
          },
          "nome": {
            "type": "string"
          },
          "genero": {
            "type": "string"
          },
          "duracao": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "diretor": {
            "type": "string"
          }
        }
      }
    }
  }
}